[CCUBE-2107][RL] Create common warning page
