+++ e2e/nextjs-app/app/components/common/index.ts
@@ -1,0 +1,1 @@
+export * from "./warning";
